ログインしているユーザーが出品者ではない場合、削除を行えない設定

+++ app/controllers/items_controller.rb
@@ -36,6 +36,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+    redirect_to root_path unless current_user.id == item.user_id
     redirect_to root_path if item.destroy
   end
